[useTableState]: separate URL storing and external value change handling from table state management into HOH(High Order Hook). Rework internal onValueChange implementation to work with updater callback.
