+++ uui-core/src/hooks/index.ts
@@ -5,6 +5,6 @@ export * from './useScrollShadows';
 export * from './useUuiError';
 export * from './useForceUpdate';
 export * from './useUuiServices';
-export * from './useTableState';
+export * from './useTableState/useTableState';
 export * from './usePrevious';
 export * from './useLayer';
